+++ app/src/main/java/com/tyganeutronics/myratecalculator/database/RTGS.kt
@@ -1,10 +1,11 @@
 package com.tyganeutronics.myratecalculator.database
 
 import com.tyganeutronics.myratecalculator.R
+import java.math.BigDecimal
 
-class RTGS constructor(id: String, rate: Double) : Currency(id, rate) {
+class RTGS constructor(id: String, rate: BigDecimal) : Currency(id, rate) {
 
-    constructor(rate: Double) : this("RTGS", rate)
+    constructor(rate: BigDecimal) : this("RTGS", rate)
 
     override fun getSign(): String {
         return "$"
