+++ app/src/main/java/com/tyganeutronics/myratecalculator/database/ZAR.kt
@@ -1,10 +1,11 @@
 package com.tyganeutronics.myratecalculator.database
 
 import com.tyganeutronics.myratecalculator.R
+import java.math.BigDecimal
 
-class ZAR constructor(id: String, rate: Double) : Currency(id, rate) {
+class ZAR constructor(id: String, rate: BigDecimal) : Currency(id, rate) {
 
-    constructor(rate: Double) : this("RAND", rate)
+    constructor(rate: BigDecimal) : this("RAND", rate)
 
     override fun getSign(): String {
         return "R"
